feat(identity): support non-PKCS#12 identities
PKCS#12 identity (aka .pfx files) are often used to transport private
key material with a decryption password.

In scenarios where a Vault-like server is used to emit PKI certificates,
PFX files are not really used and certificates private key material is
installed in non swappable RAM locations directly without any decryption
password required.

This simplifies automatic rotation and rollouts to production
deployments when this is available.

Before this, it would be necessary to re-encode PEM-encoded DER private
key via openssl with an empty password and store it in another runtime
location.

Signed-off-by: Raito Bezarius <masterancpp@gmail.com>